se separan manipulación del dom y lógica
